+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Changelog
 * Add `configuration.telemetry` to allow sending of internal errors to be disabled.
   [#1375](https://github.com/bugsnag/bugsnag-cocoa/pull/1375)
 
+* Fix data races detected by TSan in `BSGRunContextUpdateTimestamp` and `UpdateAvailableMemory`. 
+  [#1384](https://github.com/bugsnag/bugsnag-cocoa/pull/1384)
+
 ## 6.17.1 (2022-05-18)
 
 ### Bug fixes